Extract _set_callback
